Refactor Demo page into modular components
