+++ frontend/src/components/demo/ExplanationPanel.tsx
@@ -1,0 +1,121 @@
+import React, { useState, useMemo } from "react";
+import DOMPurify from "dompurify";
+
+interface ExplanationPanelProps {
+  explanation: string;
+}
+
+const ExplanationPanel = React.memo(({ explanation }: ExplanationPanelProps) => {
+  const [isExpanded, setIsExpanded] = useState(false);
+
+  const { truncatedContent, fullContent, needsTruncation } = useMemo(() => {
+    if (!explanation)
+      return {
+        truncatedContent: "",
+        fullContent: "",
+        needsTruncation: false,
+      };
+
+    const formatContent = (text: string) => {
+      return text
+        .replace(/\*\*(.*?)\*\*/g, '<h3 class="explanation-heading">$1</h3>')
+        .replace(/^- (.*$)/gim, '<li class="explanation-bullet">$1</li>')
+        .replace(/^\* (.*$)/gim, '<li class="explanation-bullet">$1</li>')
+        .replace(/^• (.*$)/gim, '<li class="explanation-bullet">$1</li>')
+        .split("\n")
+        .map((line) => {
+          const trimmed = line.trim();
+          if (trimmed.startsWith("<h3") || trimmed.startsWith("<li")) return trimmed;
+          if (trimmed) return `<p class="explanation-text">${trimmed}</p>`;
+          return "";
+        })
+        .filter(Boolean)
+        .join("");
+    };
+
+    const words = explanation.split(" ");
+    const wordLimit = 60;
+    const needsTruncation = words.length > wordLimit;
+
+    const truncatedText = needsTruncation
+      ? words.slice(0, wordLimit).join(" ") + "..."
+      : explanation;
+
+    return {
+      truncatedContent: DOMPurify.sanitize(formatContent(truncatedText)),
+      fullContent: DOMPurify.sanitize(formatContent(explanation)),
+      needsTruncation,
+    };
+  }, [explanation]);
+
+  return (
+    <div className="space-y-2 h-full overflow-y-auto">
+      <style>{`
+        .explanation-heading {
+          font-size: 16px;
+          font-weight: 700;
+          color: #1f2937;
+          margin: 8px 0 3px 0;
+          padding-bottom: 2px;
+          border-bottom: 2px solid #3b82f6;
+          display: inline-block;
+          line-height: 1.2;
+          word-wrap: break-word;
+          overflow-wrap: break-word;
+        }
+        .explanation-heading:first-child {
+          margin-top: 0;
+        }
+        .explanation-bullet {
+          list-style: none;
+          position: relative;
+          padding-left: 14px;
+          margin: 2px 0;
+          line-height: 1.3;
+          color: #374151;
+          font-size: 14px;
+          word-wrap: break-word;
+          overflow-wrap: break-word;
+          hyphens: auto;
+        }
+        .explanation-bullet:before {
+          content: "•";
+          color: #3b82f6;
+          font-weight: bold;
+          position: absolute;
+          left: 0;
+          font-size: 12px;
+        }
+        .explanation-text {
+          margin: 3px 0;
+          line-height: 1.3;
+          color: #4b5563;
+          font-size: 14px;
+          word-wrap: break-word;
+          overflow-wrap: break-word;
+          hyphens: auto;
+        }
+      `}</style>
+
+      <div
+        className="explanation-content"
+        dangerouslySetInnerHTML={{
+          __html: isExpanded ? fullContent : truncatedContent,
+        }}
+      />
+
+      {needsTruncation && (
+        <div className="mt-2">
+          <button
+            onClick={() => setIsExpanded(!isExpanded)}
+            className="text-blue-600 hover:text-blue-800 text-xs font-medium underline focus:outline-none focus:ring-1 focus:ring-blue-500 focus:ring-opacity-50 rounded"
+          >
+            {isExpanded ? "Show Less" : "Read More"}
+          </button>
+        </div>
+      )}
+    </div>
+  );
+});
+
+export default ExplanationPanel;
